+++ CHANGELOG.md
@@ -14,6 +14,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-7508](https://openlmis.atlassian.net/browse/OLMIS-7508): Add validation on product addition table.
 * [OLMIS-7537](https://openlmis.atlassian.net/browse/OLMIS-7537): Configure redux-persist
 * [OLMIS-7538](https://openlmis.atlassian.net/browse/OLMIS-7538): User can create requisition-less order while offline 
+* [OLMIS-7539](https://openlmis.atlassian.net/browse/OLMIS-7539): Offline created requisition-less orders are synchronised when user is back online 
 
 Improvements:
 * [OLMIS-7515](https://openlmis.atlassian.net/browse/OLMIS-7515): Order create table uses numeric input